pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["blyqq <f0ul2212@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.4"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
pytest = "^8.3.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
